+++ src/screens/Ticketing/Tickets/AvailableTickets/AvailableTickets.tsx
@@ -1,0 +1,70 @@
+import {View} from 'react-native';
+import Button from '@atb/components/button';
+import {TicketsTexts, useTranslation} from '@atb/translations';
+import {AddTicket} from '@atb/assets/svg/mono-icons/ticketing';
+import React from 'react';
+import {useAuthState} from '@atb/auth';
+import {StyleSheet, useTheme} from '@atb/theme';
+import {useHasEnabledMobileToken} from '@atb/mobile-token/MobileTokenContext';
+
+export const AvailableTickets = ({
+  onBuySingleTicket,
+  onBuyPeriodTicket,
+}: {
+  onBuySingleTicket: () => void;
+  onBuyPeriodTicket: () => void;
+}) => {
+  const styles = useStyles();
+  const {theme} = useTheme();
+  const {abtCustomerId, authenticationType} = useAuthState();
+  const hasEnabledMobileToken = useHasEnabledMobileToken();
+  const {t} = useTranslation();
+
+  const isSignedInAsAbtCustomer = !!abtCustomerId;
+
+  return (
+    <View style={styles.container}>
+      {isSignedInAsAbtCustomer && (
+        <View style={{padding: theme.spacings.medium}}>
+          <Button
+            mode="primary"
+            color="primary_2"
+            text={t(TicketsTexts.buyTicketsTab.button.single.text)}
+            accessibilityHint={t(
+              TicketsTexts.buyTicketsTab.button.single.a11yHint,
+            )}
+            onPress={onBuySingleTicket}
+            icon={AddTicket}
+            iconPosition={'right'}
+            testID="singleTicketBuyButton"
+          />
+          {hasEnabledMobileToken && (
+            <Button
+              mode="primary"
+              color="primary_2"
+              text={t(TicketsTexts.buyTicketsTab.button.period.text)}
+              accessibilityHint={t(
+                TicketsTexts.buyTicketsTab.button.period.a11yHint,
+              )}
+              onPress={onBuyPeriodTicket}
+              viewContainerStyle={styles.buyPeriodTicketButton}
+              icon={AddTicket}
+              iconPosition={'right'}
+              testID="periodTicketBuyButton"
+            />
+          )}
+        </View>
+      )}
+    </View>
+  );
+};
+
+const useStyles = StyleSheet.createThemeHook((theme) => ({
+  container: {
+    flex: 1,
+    backgroundColor: theme.colors.background_1.backgroundColor,
+  },
+  buyPeriodTicketButton: {
+    marginTop: theme.spacings.medium,
+  },
+}));
